refactor: split up
